+++ docs/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 For more details on new features, please check the [Manual](./MANUAL.md).
 
+<details open><summary>9.13.1 - 30 July 2023</summary>
+
+### Fixed
+- Fixed an issue with updating Extra Networks during batch processing
+
+</details>
+
 <details open><summary>9.13.0 - 30 July 2023</summary>
 
 ### About